12 support concurrent matches && 19 deck selection (#23)
* add menu stuff

* allow queuing

+++ src/Net/PlayerGameState.cs
@@ -16,9 +16,9 @@ public class PlayerGameState
         Enemy
     }
 
-    public int PeerId { get; init; }
+    public long PeerId { get; init; }
 
-    public int EnemyPeerId { get; set; }
+    public long EnemyPeerId { get; set; }
 
     public PlayerGameState Enemy { get; set; }
